bert_classifier/requirements.txt
tensorflow-text==2.8.*
tf-models-official==2.7.0


googletrans~=3.0.0
pandas
matplotlib~=3.7.0
tqdm~=4.64.1
requests~=2.28.2
tweepy~=4.12.1
nltk~=3.8.1
textstat~=0.7.3
textblob~=0.17.1
scikit-learn~=1.2.1
pysafebrowsing~=0.1.2
spacy~=3.5.0
python-decouple
tensorflow<2.10